requirements.txt
flask
logging
gunicorn
flask_cors
google-generativeai